requirements.txt
pdfkit==0.6.1
django-bulma==0.8.0

asgiref==3.4.1
bulma==1.0.1
Django==3.2.7
django-bulma==0.8.2
django-phonenumber-field==5.2.0
django-simple-bulma==2.4.0
libsass==0.21.0
pdfkit==0.6.1
phonenumbers==8.12.32
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
wkhtmltopdf==0.2
